+++ source/scener/content/gltf/buffer_view.hpp
@@ -27,10 +27,10 @@ namespace scener::content::gltf
 
     public:
         /// Gets the offset into the buffer in bytes.
-        std::size_t byte_offset() const noexcept;
+        std::uint32_t byte_offset() const noexcept;
 
         /// Gets the length of the buffer-view in bytes.
-        std::size_t byte_length() const noexcept;
+        std::uint32_t byte_length() const noexcept;
 
         /// Gets the buffer name.
         /// \returns the buffer name.
@@ -45,12 +45,12 @@ namespace scener::content::gltf
         ///               measured in bytes.
         /// \param count specifies the size in bytes of the data store region being replaced.
         /// \returns a view to the buffer data from object's data store.
-        gsl::span<const std::uint8_t> get_data(std::size_t offset, std::size_t count) const noexcept;
+        gsl::span<const std::uint8_t> get_data(std::uint32_t offset, std::uint32_t count) const noexcept;
 
     private:
         std::shared_ptr<buffer> _buffer      { nullptr };
-        std::size_t             _byte_offset { 0 };
-        std::size_t             _byte_length { 0 };
+        std::uint32_t           _byte_offset { 0 };
+        std::uint32_t           _byte_length { 0 };
         std::string             _name        { };
 
         template <typename T> friend class scener::content::readers::content_type_reader;
